Simple implementation of separation of plugins
