+++ bookmarklet.html
@@ -18,7 +18,7 @@
         Drag this button to your bookmarks bar to save it as a bookmarklet:
       </p>
       <p>
-        <a class="bookmarklet" href="javascript:(function()%7Bif(!document.getElementById(%22bookmarklet__style_bed968d%22))%7Bvar%20link%3Ddocument.createElement(%22link%22)%3Blink.id%3D%22bookmarklet__style_bed968d%22%2Clink.rel%3D%22stylesheet%22%2Clink.href%3D%22http%3A%2F%2Flocalhost%3A8080%2Fcss%2Fmain.css%22%2Cdocument.body.appendChild(link)%7Dfunction%20callback()%7B%7Dvar%20s%3Ddocument.createElement(%22script%22)%3Bs.addEventListener%3Fs.addEventListener(%22load%22%2Ccallback%2C!1)%3As.readyState%26%26(s.onreadystatechange%3Dcallback)%2Cs.src%3D%22http%3A%2F%2Flocalhost%3A8080%2Findex.js%22%2Cdocument.body.appendChild(s)%3B%7D)()">Minima</a>
+        <a class="bookmarklet" href="javascript:(function()%7Bif(!document.getElementById(%22bookmarklet__style_bed968d%22))%7Bvar%20link%3Ddocument.createElement(%22link%22)%3Blink.id%3D%22bookmarklet__style_bed968d%22%2Clink.rel%3D%22stylesheet%22%2Clink.href%3D%22http%3A%2F%2Flocalhost%3A8080%2Fcss%2Fmain.css%22%2Cdocument.body.appendChild(link)%7Dfunction%20callback()%7B%7Dvar%20s%3Ddocument.createElement(%22script%22)%3Bs.addEventListener%3Fs.addEventListener(%22load%22%2Ccallback%2C!1)%3As.readyState%26%26(s.onreadystatechange%3Dcallback)%2Cs.src%3D%22http%3A%2F%2Flocalhost%3A8080%2Fdist%2Fmain.js%22%2Cdocument.body.appendChild(s)%3B%7D)()">Minima</a>
       </p>
       
       <p class="dim">This page was created with the <a href="https://github.com/mrcoles/bookmarklet">bookmarklet</a> npm library.</p>
