+++ bookmarklet.js
@@ -2,13 +2,13 @@
 // @name Minima
 // @author mmhan
 // @style !loadOnce http://localhost:8080/css/main.css
-// @script http://localhost:8080/index.js
+// @script http://localhost:8080/dist/main.js
 // ==/Bookmarklet==
 
 // development configs
 // @style !loadOnce http://localhost:8080/css/main.css
-// @script http://localhost:8080/index.js
+// @script http://localhost:8080/dist/main.js
 
 // production configs
-// @style !loadOnce https://cdn.jsdelivr.net/gh/mmhan/minima@0.0.1/css/main.css
-// @script https://cdn.jsdelivr.net/gh/mmhan/minima@0.0.1/index.js
+// @style !loadOnce https://cdn.jsdelivr.net/gh/mmhan/minima/css/main.css
+// @script https://cdn.jsdelivr.net/gh/mmhan/minima/dist/main.js
