+++ package.json
@@ -1,13 +1,15 @@
 {
   "name": "minima-bookmarklet",
   "version": "0.0.1",
-  "main": "index.js",
   "repository": "git@github.com:mmhan/minima.git",
   "license": "MIT",
+  "private": true,
   "dependencies": {
     "bookmarklet": "^3.0.0"
   },
   "devDependencies": {
-    "http-server": "^0.12.3"
+    "http-server": "^0.12.3",
+    "webpack": "^5.33.2",
+    "webpack-cli": "^4.6.0"
   }
 }
